adapt(auth/password-reset): rest controller for password reset
