+++ libs/core/security/src/lib/authentication/presenters/http/index.ts
@@ -1,3 +1,3 @@
 export * from "./authentication";
 export * from "./authentication-token";
-export * from "./password";
+export * from './password-reset';
