+++ libs/core/security/src/lib/authentication/presenters/http/password-reset/index.ts
@@ -1,0 +1,1 @@
+export * from './password-reset.controller';
