--- libs/core/security/src/lib/authentication/presenters/http/password/index.ts
@@ -1,1 +1,0 @@
-export * from './password.controller';
